[TBTC-10] Implement `burn` variant
Problem: Right now we are inheriting the burn command from managed
ledger's interface. But the functionality is slightly different here.
Here we should be able to burn from just one account. The one that has
the address stored in the `redeemAddress` field. Thus the entry point
for burn also does not accept an address parameter.

Solution: Get the redeem address from storage and use it as the address
to burn from, and call the 'burn' entry point in managed ledger.  The
cli interface also had to be changed to not take an address parameter
for burn command.

Tests: Adds a test that test the burn entry point and checks that the
value is deducted from the balance in the address from the
'redeemAddress' field in storage. Tests for cli program also had to be
changed to reflect the change in the parameter structure.