[TBTC-12] Implement mint with tests
Problem : Right now we are inheriting the mint command from managed
ledger's interface. But the functionality is slightly different here.
Here we should only accept invocations from operators. It should also
update `totalMinted` in addition to the `totalSupply` field in storage.

Solution: Implement the endpoint such that it is callable from
operators. It should also update the `totalMinted` field in storage.