[TBTC-13] Access permissions for pause/unpause for operator/admin
Problem: The pause/unpause endpoints inherited from managed ledger
requires admin for both the entry points. But here we want to allow
operators to pause, but not unpause, and only admins to be able to
unpause.

Solution: Allow operators to pause, but only admins to unpause

Tests: Tests are included to check the required behavior.