[TBTC-14] Implement Migration using proxy agent
Problem: Make possible the migration of a contract to a newer version.

Solution: Implement the migration using the proxy method as described
in https://issues.serokell.io/issue/TM-295, thus this commit adds the
following entry points (with their inputs)

    StartMigrateTo ("migrationManager" :! MigrationManager)
    StartMigrateFrom ("migrationAgent" :! MigrationManager)
    MintForMigration ("to" :! Address, "value" :! Natural)
    Migrate ()

startMigrateTo: This entry point is only callable by admin. This input
to this entry point is `("migrationManager" :! MigrationManager)`, which
is just the address of a proxy contract of type
`ContractAddr (Address, Natural)` that will handle the conversion and
minting of tokens in the new contract, and stores it in contracts
storage, in the `migrationManager` field. Admin can call this entry
point as many times as required, and consecutive calls just overwrites
previous value with the new one. This entry point can fail with
`SenderIsNotAdmin` error.

startMigrateFrom: This entry point is only callable by admin. This input
to this entry point is `("migrationAgent" :! MigrationManager)`, which
is just the address of a proxy contract of type
`ContractAddr (Address, Natural)` that will be talking to this contract
to request tokens here on bhalf of accounts in the old contract.
minting of tokens in the new contract, and stores it in contracts. The
contract stores this address in the `migrationAgent` field. This entry
point is also callabe multiple times and each call just overwrites the
value from the earlier call with new value. This entry point can fail with
`SenderIsNotAdmin` error.

mintForMigration: This entry point is only callable by the address that
is stored in `migrationAgent` field. This mints tokens in the new
contract on behalf of accounts in the old. This entry point can fail
with `SenderIsNotAgent` error.

migrate: This entry point migrate callers tokens (all of them) in this
contract to the new contract. This

1. get the balance in the account for the sender.
2. If there is no balance, fails with `NoBalanceToMigrate` error.
3. burn all their tokens in this contract.
3. get the address of the migration manager from storage. If there is no
migration manager set, fails with `MigrationNotEnabled` error.
4. call the migration manager with the address of the sender, and their
balance.

This also adds a new module `Lorentz.Contracts.Agent` that encapsulate
the functionality of the migrationManger and can work with any contract
as long as it has a `MintForMigration` constructor with a single field
of type ("to" :! Address, "value" :! Natural).

Tests: Tests that checks all the intended behavior is included.

Dependencies: TM-299, TM-298, TM-296