[TBTC-14] Implement Migration using proxy agent
Problem: Make possible the migration of a contract to a newer version.

Solution: Implement the migration using the proxy method as described in
https://issues.serokell.io/issue/TM-295, thus this commit adds the
following entry points (with their inputs)

    StartMigrateTo ("migrationManager" :! MigrationManager)
    StartMigrateFrom ("migrationManager" :! MigrationManager)
    MintForMigration ("to" :! Address, "value" :! Natural) Migrate ()

This entry point is only callable by admin. This input to this entry
point is `("migrationManager" :! MigrationManager)`, which is just the
address of a proxy contract of type `ContractAddr (Address, Natural)`,
that will handle the creation of target contract parameter to
trigger the minting of migrated tokens in the new contract. This
address is stored in contract's storage, in the
`migrationManagerOut` field. Admin can call this entry point as
many times as required, and consecutive calls just overwrites
previous value with the new one. This entry point can fail with
`SenderIsNotAdmin` error.

This entry point is only callable by admin. This input to this entry
point is `("migrationManager" :! MigrationManager)`, which is just the
address of a proxy contract of type `ContractAddr (Address, Natural)`
that will be talking to this new contract to request tokens to be minted
here on behalf of accounts in the old contract.  The contract stores
this address in the `migrationManagerIn` field. This entry point is also
callable multiple times and each call just overwrites the value from the
earlier call with new value. This entry point can fail with
`SenderIsNotAdmin` error.

This entry point is only callable by the address that is stored in
`migrationManagerIn` field.  This mints tokens in the new contract on
behalf of accounts in the old. This first check if the
`MigrationManagerIn` field is set, and if it is not set, fails with
`MigrationNotEnabled` error. If it is available, it checks if the value
it contains is the same as the sender's address. If it is not then the
contract fails with `SenderIsNotAgent` error. If both checks pass, then
the contract calls the `mint` procedure to mint tokens for the account
along with the associated bookkeeping.

This entry point migrate callers tokens (all of them) in this contract
to the new contract. This

1. Gets the balance in the account for the sender.

2. If there is no balance, fails with `NoBalanceToMigrate` error.

3. Burn all their tokens in this contract.

4. Get the address of the migration manager from storage. If there is
no migration manager set, fails with `MigrationNotEnabled` error.

5. Call the migration manager with the address of the sender, and
their balance.

This is an end user call that is only possible when the contract is in a
running (unpaused) state.

This PR also adds a new module `Lorentz.Contracts.Agent` that
encapsulates the functionality of a migration manger and can work with
any contract as long as it has a `MintForMigration` constructor with a
single field of type `("to" :! Address, "value" :! Natural)`.

So the process of migration is as follows.

1. Admin deploys the migration agent passing the addresses of the old
contract and the new.

2. Admin/Operator pauses the old contract.

3.  Admin calls Contract1's `startMigrateTo` entry point passing the address
of the migration agent.

4. Admin calls Contract2's `startMigrateFrom` entry point passing the
address of the migration agent.

5. Admin unpauses Contract1.

Tests: Tests that checks all the intended behavior is included.

Dependencies: TM-299, TM-298, TM-296