[TBTC-14] Ensure paused state in `startMigrateTo` call
Problem : Right now the contract allow a call to `startMigrateTo`
when the contract is in a running state. This is contrary to what the
requirement says. Also it might result in funds being lost, because
a `migrate` call to the first version might ends up burning the tokens
without minting the same in new version if the `startMigrateFrom` call
is still unfinished in the new version.

Solution: Ensure that the contract is paused, and no end user calls
including `migrate` calls, is processed when the migration is in
progress.

Tests: Two tests were added that test that call to `startMigrateTo`
is denied if the contract is in a running state, and one that tests
`migrate` calls are denied if the contract is in a paused state.