[TBTC-20] Add missing fields to contract storage
Problem: Contract storage does not have tokenname and code fields

Solution: Add the `tokenname` and `code` fields to storage and populate
them while initialising the contract. Right now `tokenname` is set to
`TZBTC` and code to `ZBTC`.

Tests: The bats test for `printInitialStorage` command was modified to
reflect the change.