[TBTC-9] Signing with `tezos-client`
Problem: Signing is complicated. Moreover, later we want to support
ledger. So, signing is definetly not going to happen in RPC client.

Solution: For now, we decided to use `tezos-client` for operation
signing. Now subcommands of `tzbtc-client` return unsigned operation
in hexademical representation. Also new subcommand `injectOperation`
was added. For more convenient work with `tzbtc-client` and
`tezos-client` script `/scripts/tzbtc.sh` was added. It has
interface similar to `tzbtc-client` but it submits transactions
to the chain by geting unsigned operation, signing it with the
`tezos-client` and then injecting it to the chain via `tzbtc-client`.