[TBTC-17] Multisig operations and supporting infra
Problem: We need a way to have multi-sig admins and operators

Solution: We decided to go with the generic multisig contract available
at [1]. The non generic contract that is available at the moment is only
able to call a contract with a unit parameter, and the TZBTC contract
did not have a unit parameter. So we needed the generic multisig
contract , which could accept a lambda in it's payload, and after
validating the signatures, could execute the lambda.

Workflow with three signers Alice, Bob, Carlos

1. Originate the mutli-sig contract. Initial storage contains

    1. Public keys of Alice, Bob and Carlos 2. Desired threshold, ie
    minimum number of valid signatures required for executing the
    operation.

2. In the main contract, either set the administrator to the address of
the multi-sig contract, or add the address of the multi-sig contract as
one of the operators.

Executing an action

1. Let Alice be the one who wants to run the operation. She uses the
`mkPackage` function to create a value of type `Package` and writes it
to a file using the `encodePackage` function. (ofcourse she will not use
these directly, but will just call a cmd line to execute these actions)
The `Package` type looks something like this.

``` data Package =
      Package
        { pkToSign :: !Text
        , pkUnpackEnv :: !TcOriginatedContracts
        , pkSignatures :: ![(PublicKey, Signature)]
        }
```
The signature field is a list of tuples with each containing a public
key and a signature. We are collecting public key because when a request
is submitted with a list of signatures, the contract expects the
signature list to be in the same order that the contract have them in
it's storage. So at the time of submission, the caller provides a master
list of public keys. The signatures from the packages will be used to
fill a list in the same order as the master list, filling with Nothing
value for public keys for which a signature was not available from all
the packages.

2. Then she uses the `getBytesToSign` function to get the byte sequence
that needs to be signed. She makes a signature out of the byte sequence,
and uses the `addSignature` function to add this signature  and
public key, to the package file.

3. Now she sends this file to both `Bob` and `Carlos`. They call the
`getOpDescription` function to check the contents of the package and
then follow the procedure in step 2, and adds their signatures to the
package files. They send the new package files to Alice.

4. Alice now has three package files. Now she can call the
`mkMultiSigParam` function passing all the package files she has, and
the signatures in them will be combined, and will be forged into a
parameter for the multi-sig contract. Calling the contract with this
parameter will execute the lambda and thus call the action in the main
contract. (Actually calling the `mkMultiSigParam` also require a list of
all public keys for the signers, sorted in the same order in which
they are stored in the contract's storage).

This PR, for the time being does not solve how the list of master public
keys is made available at the time of submission. It can either be done
by maintaining all the public keys for all the signers via a
configuration file, or by looking into the contract's storage at the
time of submission.

Tests: An integration test that test the happy path of the multi-sig
workflow has been added.

[1] https://github.com/murbard/smart-contracts/blob/master/multisig/
    michelson/generic.tz