[TBTC-19] Reduce gas costs
Problem: Since migrations involve two copies of the main contract and a
migration manager, the whole process, when the end user calls the
`migrate` entrypoint, cannot fit within the hard gas limit of 800k.

Solution: Convert the storage of the contract to use the UStore infra
which enables storage of multiple bigmaps and store some of the
entrypoints inside the bigmaps in a packed form. The whole solution
mainly involves the following changes.

1. Change the storage to use a `UStore` instead of just a Bigmap as it
was the case earlier. The recent changes in mananged ledger allowed
using it with storage variants such as these.

2. Aggregate some of the entry points to a single entry point in the
main contract's interface. Then pack the handler for this aggregate
entry point in the bigmap. This was so that we can save duplication of
code to load and execute entrypoints from storage to a single instance,
as opposed to repeating it for every entrypoint that has to be
loaded from the stored form. This means that every call to an entry
point inside the aggregated entrypoints is costs more than the
current implementation, and the implementation that repeats
load/execute calls for every aggregate entry point. But since we
have left the migration entrypoints out of the aggregation, the
migration process might fit in the gas limits.

Tests: All the tests were changed to accomodate the changed
storage/parameter.