[TBTC-18] Direct mulstisig storage extraction
Problem: We can't convert raw multisig storage to the haskell value
for now because public keys are represented as a strings (our unpacker
can deal only with the public keys represented as a byte sequences).

Solution: We still can pattern match on specific shape of `Expression`
that is supposed to be the multisig storage and thus extract counter,
threshold and public keys.