[TBTC-25] Replace `contract p` with `address` in storage
Problem: It appears that in Babylon, we will no longer be store values
of type contract in storage [1]. Right now we have contract values in
storage for migration manager address. We should change this to address
and the code that uses this field accordingly.

Solution: Replace `contract p` values in storage with `address` values
and convert them to contract in the required places in contract code.

Tests: Tests were changed to build with the change.

[1] https://tezos.gitlab.io/mainnet/protocols/005_babylon.html