[TBTC-33] Make `setupClient` arguments optional
Problem: Right now the command accepts/requires all the configuration
values in one go. But this might not be the best user experience. We can
make all the inputs optional. If an input is absent in the CLI call, we
just create the config file by using a placeholder help text in its
place. And in the CLI output, we just indicate that the config file is
incomplete, as a warning.

Solution: The following changes are part of this PR

1. If the input arguments are absent, write the configuration file with
place holders.

2. If the configuration file exists, already, then will not overwrite
it. Even if the required arguments are present. The user will have to
remove the file manually.

3. Print the config file path after the action.

4. Add a `config` command that would print the config file path and
contents, when called on it's own, or would edit the config file in
place when combined with `--edit` switch and config values that needs
changing.