[TBTC-35] Human readable package descriptions.
Problem : We have package files, but when a user wants to sign the
contents, there is no easy way for them to check what is in it. The
output of `getOpDescription` operation is not helpful as of now.

Solution:
As a solution, store the original parameters ie, (main contract address,
param, counter value) in a packed form inside the multi-sig file.
During the `getOpDescription` call, use the packed params  to obtain the
original parameters. Check that it matches with the bytes that are being
signed. If they match then display it to the user, or else show an
error.