[TBTC-51] Config edit via external editor
Problem : We have a config command with --edit flag that lets user edit
various config values by providing them in the command line. But it
would be better if we can open an external editor to let the config
values more easily. A problem here is that the concept of default editor
only exist in Linux like operating system, and not on windows. The lib
we are using to open the default editor, was also not available in
Windows.

Solution : Use conditional dependencies to detect os and include the
`editor-open` package. Use conditional pragmas in code to print out a
'not available on windows' error upon the invocation of this command.

In addition to this, this PR implements a number of changes so that the
config editing via `--edit` switch also works well together with the
`setupClient` command behavior. Basically an option to read config as
raw text values was added, and the `--edit` operation now reads/write
raw text values instead of the actual field types. The user input is
parsed properly in as part of the command, so there is less chance that
a bad value could end up in the config. The current behavior can be
summarised as follows.

1. At the start user invokes `setupClient` command. If the required
values are missing, then those fields will be replaced with a
placeholder in the generated config file. After this user will be able
to set individual fields using the `--edit` switch. This is possible
even with the presence of placeholders in the file. After editing, the
config file is checked for validity, and if found invalid, a warning is
printed out.

Tests: Couple of tests have been included that test that the placeholder
values are properly generated, and while reading these placeholders are
processed properly.