[TBTC-45] Address preserving migrations
Problem : We have to implement migrations that does not change the
address of the contract.

Solution: Use the upgradeable architecture and infra to change the
contract implementation to an entrypoint wise upgradeable one. This
means that we start with a contract with a mostly empty storage, and
upgrade it to a fully functional one in the first upgradation step.

Most of the functionality is implement in the two new V0 and V1 modules
The Rest of the changes are,

1. Refactoring of the contract parameter.

Earlier the contract parameter was in the form of two top level
constructors. ParameterWithViews and ParameterWithoutViews. But it
appeared that we only need one constructor to aggregate all the 'safe'
entrypoints, that does not have things like contract p values in its
arguments. So the parameter is now of the following form, where we have
all the 'unsafe' entrypoints at top level itself, and the 'safe' ones
wrapped under the SafeEntrypoints constructor.

data Parameter (interface :: [EntryPointKind])
  = GetVersion (View () Natural)
  -- TZBTC Entrypoints
  | GetAllowance        !(View ManagedLedger.GetAllowanceParams Natural)
  | GetBalance          !(View Address Natural)
  | GetTotalSupply      !(View () Natural)
  | GetTotalMinted      !(View () Natural)
  | GetTotalBurned      !(View () Natural)
  | GetAdministrator    !(View () Address)
  | SafeEntrypoints (SafeParameter interface)

data SafeParameter (interface :: [EntryPointKind])
  = Run (UParam interface)
  | Upgrade UpgradeParameters
  | SetMaster Address
  | EpwBeginUpgrade Natural  -- version
  | EpwApplyMigration ("migrationscript" :! MigrationScript)
  | EpwSetCode ("contractcode" :! ContractCode)
  | EpwFinishUpgrade
  | Transfer            !ManagedLedger.TransferParams
  | TransferViaProxy    !TransferViaProxyParams
  | Approve             !ManagedLedger.ApproveParams
  | ApproveViaProxy     !ApproveViaProxyParams
  | Mint                !ManagedLedger.MintParams
  | Burn                !BurnParams
  | AddOperator         !OperatorParams
  | RemoveOperator      !OperatorParams
  | SetRedeemAddress    !SetRedeemAddressParams
  | Pause               !()
  | Unpause             !()
  | TransferOwnership   !TransferOwnershipParams
  | AcceptOwnership     !AcceptOwnershipParams
  | SetProxy            !SetProxyParams

Tests were also changed to accomodate the new origination procedure.
Some new tests were added to check the migration specific entrypoints
and statuses.