[TM-342] Update scripts to use babylonnet
Problem: we have some scripts for testing and they are
using hardcoded `alphanet.sh` constant and refer to alphanet in
variable names and comments.
But alphanet is about to be decommissioned and babylonnet will be
the primary network for testing instead.

Solution: simply replace all occurrences of alphanet with babylonnet
as it is done in `morley` repo.