[TBTC-58] Remove Proxy and related stuff
Problem: we have proxy contract that is needed for FA1.2.1
compatibility which is useless in Babylon. So it is essentially
just some useless code which also makes the contract bigger.
Solution:
* remove proxy contract;
* remove `*ViaProxy` and `SetProxy` entrypoints;
* remove proxy from the storage;
* remove tests for proxy.

I have removed mentions of both proxy and agents from README.
Agent contract is still present, but will be removed as part of
TBTC-45.