[TBTC-55] Update storage getters.
Problem: New upgradability approach arrived. Now we have UStore as a
storage. This means that fields (unrelated to the upgradability) are now
stored in `BigMap ByteString ByteString`.

Solution: Update getters, so that they now pack the field name (if we
are getting the UStore field) or pair of submap name and key (if we
are getting value by key from UStore submap) perform get from BigMap
using RPC call, and unpack the value back to the desired type.