[TBTC-54] Deploy contract via RPC
Problem: With new upgradeability approach contract deploying became
quite complex. In order to deploy the contract we have to originate it
and perform a bunch of transactions in order to upgrade the contract
to V1. It is inconvenient to make it manually, it is also inconvenient
to use python script.

Solution: Implement contract deploying via RPC. The contract origination
via RPC is implemented. Upgrade is implemented using existing
`runTransaction` function.