[TBTC-62] Merge `admin` and `master` fields
Problem : Right now the contract has a master field, which is part of the concrete
storage, and an admin field, which is part of the wrapped ustore. master
dictates who can do the migration, where as admin controls the access to
the domain logic. We have to merge these two into a single field.

Solution : Eliminate the master field, and only have admin
field inside ustore. We will have to change the v0 version to always
have this field in the wrapped ustore. And the upgradeability entry
points will read from this field from ustore to access control upgrade
operations.

Tests: Bats tests were modified to include the changed initial storage
in the output of command that prints initial storage.