[TBTC-70] Fix multisig call
Problem: When a multisig action is called using a package with
signatures, the client errors out with a wrong parameter error. This
seems to be happening because of the new entry point feature of babylon.
Since the multisig contract has a default entrypoint, the when the call
is made without specifying an entrypoint, it gets routed to the default
entrypoint which expects a unit value.

Solution: It seems that the RPC api already provides a way to specify
the entrypoint. Use it to call the main entrypoint of multisig contract
during multisig call.