[TBTC-73] Make documentation link relative
Problem: documentation link is absolute and does not work in forks.

Solution: make it relative.
This solution is BAD and temporary.
The link will only work at the home page of the repo.
We should revert it once people get direct access to the repo.
Also it breaks crossref-verifier, so we ignore README.md for now.