[TBTC-79] Make tokenCode and tokenName configurable
Problem: currently these values are always hardcoded to "TZBTC", it
is not possible to set them to anything else using our CLI w/o
manual modification.

Solution:
1. Pass these values in CLI.
2. Pass full `OriginationParameters` to `deployTzbtcContract`
3. Rename opMaster to opOwner for consistency since related code was
touched anyway.
4. There used to be `originationParams` function with hardcoded
meta data. It's not useful in production anymore (because now all
values are configurable), so it was removed from the library.
Such a function still makes sense in tests, so there is
`dummyOriginationParameters` which does the same thing.