[TBTC-79] Rename and document tokenname and tokencode
Problem: `tokenname` and `tokencode` do not follow our common
naming convention or nearly any other convention (there are
two words in each name and they are not visually distinct).
Also these fields are not documented.

Solution:
1. Rename them to camlCase style.
2. Add Haddock comments.

Note: it changes the corresponding UStore keys, but it should be fine
because we have renamed administrator to owner recently anyway, so
one key has changed.