[TBTC-79] Add Get{TokenName,TokenCode,RedeemAddress} entrypoints
Problem: we store these things, but they are not exposed by
any entrypoint. One can get the only from the storage, but
it is inconvenient and unstable.

Solution: add `GetRedeemAddress`, `GetTokenName` and `GetTokenCode`
entrypoints. They are implemented trivially, the same way as
existing such getters.
I generalized `getTotal` a bit to make it work with a field of
any type.

It is not obvious whether they should be void or view entrypoints.
I have made them view entrypoints because all other getters are
views, so at least the users of the contract don't have to deal with
another mechanism.