Add tests on generated documentation
Problem: practice has shown that it's easy to accidentially break
autodoc so that some entrypoints descriptions are missing, or properties
are gone.
Morley now contains some test predicates on autodoc, would be nice to
include them here.

Solution: add Morley's common tests on autodoc.

Note that currently we have an issue related to the fact that some code
is called from storage and thus its properties are not reflected in
the generated documentation; ATM Morley does not contain tests covering
this case, but it should after we gather more info about similar
accidents.