[TBTC-85] Add global `--user` option to override default alias
Problem: We use the `tezos-client` alias `tbtc-user` by default, for
sending operations. But sometimes we might need to send transactions as
a different user. Right now this is only possible by re-aliasing the
`tzbtc-user` alias to the new address. We need to add a way to override
the default `tzbtc-user` alias for all commands that send network
operations

Solution: The PR adds a global option `--user` that accept an address or
alias and use it to override the default `tzbtc-user` alias to send
network operations.

Internally, this PR replaces the IO monad with a `ReaderT AppEnv IO`
where we store the user override information from the command line
argument. The `HasConfig` method looks up the override from the
environment of the reader, and adjust the value of `userAlias`
configuration values accordingly. So the code of the individual commands
remains the same, and the changes are restricted mostly to top level
code.

Tests: A test has been added in the test suite that checks the output
of config command contains the overridden aliased user. Some Bats tests
were also added to check the existence of `--user` argument.