[TM-381] Use specialized multisig
Problem: We have to use specialized multisig, because it restricts
possible operations that can be triggered by the multisig to the subset
of operations of the target contract only.

Solution: We have added a specialized multisig contract to
`morely-multisig` package. We use that contract here. The way it will
work is as follows. When deploying the multisig contract, we generate a
multisig contract that is specialized to the parameter/entrypoint,
"safeEntrypoints" of the TZBTC contract. To make this work, we slightly
change the parameter of TZBTC contract to annotate the `SafeEntrypoints`
branch in the original parameter (So that it can be called by a named
entrypoint call by the multisig). This is done by using a combination of
two `ParameterHasEntryPoints` instances implemented via `EpdRecursive`
and `EpdDelegate` for main `Parameter` and `SafeParameter` types
respectively.