[TBTC-92] Test entrypoint structure
Problem: We need a test to check all the expected entrypoints are
available, with their expected type. This will help us notice if
something changes in the entrypoint structure.

Solution: Implement this test. We have placed the reference parameter as
a Michelson contract that has the parameter of TZBTC contract with empty
storage and body. In the test, we parse this contract into an untyped
contract. Then we walk the parameter of this contract, making a map of
entrypoint name to the types at each entrypoint. We then compare this
with the real entrypoints from the real tzbtc parameter. A mismatch is
thrown as test failure.