[TBTC-94] Option to override multisig alias
Problem: Right now tzbtc-client by default uses the alias
`tzbtc-multisig` to get the multisig contract address while creating a
multisig package. But sometimes, we might have to work with more then
one multisig contract, say one for an operator, and another one for an
administrator. There is no easy way to handle this case. User have to
keep renaming the contract they want to use to tzbtc-multisig, before
creating the multisig package.

Solution: Use the config override infra to add this override, and change
`readConfig` function to take it into account while creating the global
config for the application.