[TBTC-86] Use morley-nettest for smoke tests
Problem: We have a small smoke test scenario in
`Lorentz.Contracts.TZBTC.Test`. It uses Lorentz.TestScenario
functionality and relies on some python scripts.  This stuff is fragile
and deprecated. There is a more modern library for that: morley-nettest
(part of morley). So we should use it instead.

Solution: Change the smoke tests in `Lorentz.Contracts.TZBTC.Test`
module to use infra from `morley-nettest`.