[TBTC-86] Use 'tzbtc-client' to run the tests
Problem: We want to use `tzbtc-client` instead of `tezos-client` to
run the tests. But the interface is slightly different.

Solution: Here we have implemented two method of the `NettestImpl`
record, `niTransfer` and `niOriginate`, via the `tzbtc-client` and
delegate the rest of the method to the `tezos-client` itself.

The `niOriginate` throws away most of the origination data and only use
`odFrom` to send the origination transaction, while originating the
`tzbtc` contract itself. The return address is parsed from the output of
`tzbtc-client` and returned.

For the `niTransfer` method, we check the passed in parameter value,
after an `toVal`/`fromVal` round trip on it, with a `Typeable` casting,
in between, we ends up with a possible `tzbtc` parameter. If we didn't
get a `tzbtc` parameter, then the `transfer` operation is passed on to
`tezos-client` itself. It we have a `tzbtc` parameter, we look at the
operation, and invokes the corresponding `tzbtc-client` command.

To make this possible, an override to the default tzbtc contract alias
was also added to the `tzbtc-client` program. This was required so that
the nettest created contract could be specified for `tzbtc` contract
operations.