[TBTC-99] Add global '--fee' option
Problem: Right now we calculate the baker fee based on operation size.
But sometimes user might want to specify an explicit fee. Like a higher
one so that the transactions are prioritized by the bakers.

Solution: Add a global `--fee` argument to explicitly set the fees for
all the transactions that are done as part of a command.