[TZBTC-89] Enable roundtrip test
Problem: roundtrip test for upgrade parameters has been disabled. It
does not work because `PermanentImpl` is an extended lambda inside, and
`fromVal . toVal` roundtrip effectively adds `PAIR; UNPAIR` prefix to
it.

Solution: it looks like the initial assumption on that `pack/unpack`
roundtrip should produce a code equal to the original one is wrong.
It worked because `==` on instructions does not check that instruction
tree structure is the same, and it also ignores `DOC_ITEM`s; however,
soon it won't be so in Morley.

Herewith, it seems reasonable to require not `unpack . pack = pure`,
but `optimize . unpack . pack = optimize` in our roundtrip test.
Later, `==` should also be replaced with an equality checking function
which is tolerant to semantically innoticable changes.

Also, in our test I now check roundtrip not for `SafeParameter`, but
only for parameter of its `Upgrade` entrypoint - that eases debugging
via checking roundtrip property for separate pieces of upgrade
parameters.