[TBTC-99] Refactor tezos-client infra, add verbose flag
Problem: There was a lot of code duplication and issues with handling
of the errors in the tezos client infra.

Solution: Refactor to remove code duplication and fix issues with error
handling. Also, to make debugging easier and bug reporting more detailed
add a `--verbose` global switch to turn on verbose logging, which will
print out all `tezos-client` interactions.