[TBTC-103] Update links
Problem: the repo is being moved to another github organization,
some links will become invalid because of that.

Solution: replace `serokell/tezos-btc` with `tz-wrapped/tezos-btc`.
Also set `github` field in `hpack` instead of setting `git`,
`bug-reports`. It's a feature of `hpack`, there is no `github` field
in `.cabal` files.
`homepage` was updated to https://tzbtc.io/
`bug-reports` now points to github because the repo is going to become
public.