[#121] Fix loss of precision in burn-cap roundtripping.
Problem: We have been using `Double` type to hold the burn-cap value
parsed out from the output of `tezos-client` program. But this caused
slight loss of precision which ended up causing this bug.

Solution: Switch to using `Scientific` instead of `Double` to hold
`burn-cap` and `baker-fee` values. Add a test to check the round trip.