[#129] Fix nettest scenario
Problem: the test scenario got broken, tezos-client complains that we
ask to send transactions from unknown address aliases.

Solution: this happens because in some places we attach a prefix to
aliases, and in some we don't. Probably that is because our scenario
uses hybrid of Morley's nettest and tzbtc's custom primitives.

I tried to use addresses instead of aliases, but it looks like
tezos-client does not know how to reveal public keys of addresses, it
can accept only aliases, and we need revealing functionality.

Eventually I started prefixing all aliases sort of manually at the top
level of the scenario. While this solution is not the best, it is
temporary so we can probably live with this for a while.