[#129] Use TokenMetadata from `morley-ledgers`
Problem: `morley-ledgers` package now has `TokenMetadata`, and it is
cleaner and more compliant (in annotations) than our `TokenMetadata`.

Solution: replace our `TokenMetadata` type.

I also had to update `parseSingleTokenMetadata` from CLI, and to be able
to rewrite it with `ApplicativeDo` I moved it to CLI module (the
previous mix of Haskell and Lorentz code in one module seemed pretty
wrong).