[#129] Use older version of `approvableLedgerSpec`
Problem: we use tests from ManagedLedger and they have gained more
suites checking e.g. that if `A` transfers from `B` to `B`, that spends
allowance. Using the old version of ManagedLedger, we don't have this
behaviour (for us loop transfer is always a no-op).

Solution: module which exports `approvableLedgerSpec` also exports
building blocks for the old functionality, we are starting to use them
here.