Fix link to FA1.2
Problem: that links points to master, and after recent (or actually not
too recent) reorganization in TZIP repository the link got outdated.
Thanks to GitLab recent features, it is not easily possible to detect
such broken links automatically.

Solution: fix the link.

Even if FA1.2 is ever updated, we comply with a particular
version of FA1.2, so let's provide a link to specific commit.