[#139] Increase precision of burn cap
Problem: we print always burn cap with precision set to 3.
It used to work before Delphi where the smallest possible
burn cap was divided by 4. So in some cases we print smaller
burn cap than required.

Solution: increase precision to 6.
It's a temporary solution, we should switch to morley-client
instead and we will do it soon.
For this reason I am not adding any tests.
Hardcoding this precision doesn't sound like a good long term solution
btw, but again, it's temporary.
The same issue is present in the existing tests, so we fix it the same
way.