Refactor nettest running
Problem: Currently emulator and nettest tests are run together, thus we
run emulator tests twice on 007 and 008 job. Which is probably not a big
deal, but still not completely proper.

Solution: Run emulator tests and nettests in separate job by using tasty
pattern to filter only jobs that run on the real network.