[#160] Examine bigmaps in non-emulated nettests
Problem: some tests run only on emulated environment due to using
`getFullStorage`. We can rewrite those to run on the real chain.

Solution: working with UStore from Haskell is a little tricky, so
I wrote a helper 'readUStore' which gets an element from UStore given
a 'UStoreElemRef' and a big map id. The rest is pretty straightforward.