[Chore] Generate unique moneybag for each test run
Problem: Since Ithaca it's quite hard to workaround various
race-conditions from the morley-client side. Moreover, adding
workarounds to it isn't really proper because these race conditions are
actually caused by the CI setup not by the library or Tezos itself.

Solution: For each job that performs tests on a real chain (either
a local-chain or a testnet) generate a unique moneybag address and
transfer some amount of money to it.