[#173] Use refillable addresses everywhere
Problem: tests tend to error out with insufficient balance errors. We
hard-code some amounts in tests, but this is error-prone and not always
sufficient.

Solution: Use refillable addresses everywhere. This is basically free
for cases that don't need refills, and will mostlysave us from running
out of funds mid-test.

One particular hard-coded transfer has to be left in place because
that's running through tzbtc-client, and tzbtc-client doesn't do
refills.